requirements.txt
pandas>=2.0.0
selenium
webdriver-manager
thefuzz
scikit-learn
nltk
streamlit
transformers
torch
plotly
pytest
fpdf2
google-generativeai>=0.8.0
python-dotenv